# [WIP] Makevalid optimizations