--- a/planar/intersect/ring.go
+++ b/planar/intersect/ring.go
@@ -17,9 +17,16 @@ type Ring struct {
 }
 
 func NewRing(segs []geom.Line) *Ring {
+	var index *SearchSegmentIdxs
+	if len(segs) > len(staticIdxs) {
+		// Only build index for large rings. Cost of building the index is
+		// higher then the query efficiency for smaller rings.
+		// TODO: re-evaluate cut-off if index implementation changes.
+		index = NewSearchSegmentIdxs(segs)
+	}
 	r := &Ring{
 		segs:  segs,
-		index: NewSearchSegmentIdxs(segs),
+		index: index,
 		bbox:  new(geom.Extent),
 	}
 	for i := range segs {
@@ -57,8 +64,12 @@ func (r *Ring) Extent() *geom.Extent {
 	return r.bbox
 }
 
+// Static indices for Ring.ContainsPoint. We build our result slice from this
+// array to avoid allocation of a new []int slice for each ContainsPoint call.
+var staticIdxs = [...]int{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15}
+
 func (r *Ring) ContainsPoint(pt [2]float64) bool {
-	if r == nil || r.index == nil {
+	if r == nil {
 		return false
 	}
 	if debug {
@@ -72,7 +83,13 @@ func (r *Ring) ContainsPoint(pt [2]float64) bool {
 	}
 
 	l := geom.Line{{r.bbox.MinX() - 1, pt[1]}, pt}
-	results := r.index.SearchIntersectIdxs(l)
+
+	var results []int
+	if r.index != nil {
+		results = r.index.SearchIntersectIdxs(l)
+	} else {
+		results = staticIdxs[:len(r.segs)]
+	}
 
 	if debug {
 		log.Printf("\t SearchIntersect got back (%v):  %+v", len(results), results)