--- a/planar/intersect/ring.go
+++ b/planar/intersect/ring.go
@@ -13,7 +13,7 @@ type Ring struct {
 	index         *SearchSegmentIdxs
 	IncludeBorder bool
 
-	bbox *geom.Extent
+	bbox geom.Extent
 }
 
 func NewRing(segs []geom.Line) *Ring {
@@ -27,7 +27,6 @@ func NewRing(segs []geom.Line) *Ring {
 	r := &Ring{
 		segs:  segs,
 		index: index,
-		bbox:  new(geom.Extent),
 	}
 	for i := range segs {
 		r.bbox.AddPoints(segs[i][0], segs[i][1])
@@ -36,7 +35,7 @@ func NewRing(segs []geom.Line) *Ring {
 }
 
 func NewRingFromPointers(pts ...geom.Pointer) *Ring {
-	var segs []geom.Line
+	segs := make([]geom.Line, 0, len(pts))
 	lp := len(pts) - 1
 	for i := range pts {
 		xy := pts[i].XY()
@@ -48,7 +47,7 @@ func NewRingFromPointers(pts ...geom.Pointer) *Ring {
 }
 
 func NewRingFromPoints(pts ...[2]float64) *Ring {
-	var segs []geom.Line
+	segs := make([]geom.Line, 0, len(pts))
 	lp := len(pts) - 1
 	for i := range pts {
 		segs = append(segs, geom.Line{pts[lp], pts[i]})
@@ -61,7 +60,7 @@ func (r *Ring) Extent() *geom.Extent {
 	if r == nil {
 		return nil
 	}
-	return r.bbox
+	return &r.bbox
 }
 
 // Static indices for Ring.ContainsPoint. We build our result slice from this