--- a/planar/makevalid/tegola/makevalid_test.go
+++ b/planar/makevalid/tegola/makevalid_test.go
@@ -11,6 +11,14 @@ import (
 )
 
 func TestMakeValid(t *testing.T) {
+	checkMakeValid(t)
+}
+
+func BenchmarkMakeValid(b *testing.B) {
+	checkMakeValid(b)
+}
+
+func checkMakeValid(tb testing.TB) {
 	type tcase struct {
 		MultiPolygon         *geom.MultiPolygon
 		ExpectedMultiPolygon *geom.MultiPolygon
@@ -19,7 +27,7 @@ func TestMakeValid(t *testing.T) {
 		didClip              bool
 	}
 
-	fn := func(t *testing.T, tc tcase) {
+	fn := func(t testing.TB, tc tcase) {
 		hm, err := hitmap.NewFromPolygons(tc.ClipBox, tc.MultiPolygon.Polygons()...)
 		if err != nil {
 			panic("Was not expecting the hitmap to return error.")
@@ -61,6 +69,15 @@ func TestMakeValid(t *testing.T) {
 	}
 	for name, tc := range tests {
 		tc := tc
-		t.Run(name, func(t *testing.T) { fn(t, tc) })
+		if t, ok := tb.(*testing.T); ok {
+			t.Run(name, func(t *testing.T) { fn(t, tc) })
+		} else if b, ok := tb.(*testing.B); ok {
+			b.Run(name, func(b *testing.B) {
+				b.ReportAllocs()
+				for i := 0; i < b.N; i++ {
+					fn(b, tc)
+				}
+			})
+		}
 	}
 }