Do not merge! This is pull request to discuss some improvement. Feel free to cherry pick single changes.