@olt 2018-08-09
math.Big showed up in my first profiling. 
I was able to change the LineIntersectBigFloat function to use float64 and the tests are still passing (except minor rounding differences in the output).

I implemented the check if the intersection point is on both line segments inside the LineIntersect function (using t and u from the formula on wikipedia).