@gdey 2018-08-09
@olt There is a LineIntersect right above (https://github.com/go-spatial/geom/blob/edf63b58dbba78c2ea03221f7ddf91388845a984/planar/line_intersect.go#L47) that uses Float64, I was using the LineIntersectBigFloat because I was getting rendering issues with the older algorithm. Just forgot to switch it out as I'm still seeing some rendering issues, but I believe these are not due to Precision as with the older Algo.