@olt 2018-08-13
I'm guessing that using `t` and `u` directly is more robust. 

Regarding rendering issues, what is the best way to share rendering issues. I can share a minimal GeoPackage with Tegola configs and tile coordinate for `xyz2svg`.