@olt 2018-08-15
I've added a few more optimizations. See my review comments for details.

This gives an additional boost (compared to the first LineIntersectFloat changes):

```
benchmark                                                                        old ns/op     new ns/op     delta
BenchmarkMakeValid/makevalidTestCases_#3_Square-4                                40474         33179         -18.02%
BenchmarkMakeValid/makevalidTestCases_#0_Triangle-4                              29651         25593         -13.69%
BenchmarkMakeValid/makevalidTestCases_#1_Four_Square_IO_OI-4                     84568         69864         -17.39%
BenchmarkMakeValid/makevalidTestCases_#2_Four_columns_invalid_multipolygon-4     168397        141918        -15.72%

benchmark                                                                        old allocs     new allocs     delta
BenchmarkMakeValid/makevalidTestCases_#3_Square-4                                401            298            -25.69%
BenchmarkMakeValid/makevalidTestCases_#0_Triangle-4                              303            237            -21.78%
BenchmarkMakeValid/makevalidTestCases_#1_Four_Square_IO_OI-4                     788            565            -28.30%
BenchmarkMakeValid/makevalidTestCases_#2_Four_columns_invalid_multipolygon-4     1413           1029           -27.18%
```


Compared to the current BigFloat implementaiton:

```
benchmark                                                                        old ns/op     new ns/op     delta
BenchmarkMakeValid/makevalidTestCases_#1_Four_Square_IO_OI-4                     188019        69864         -62.84%
BenchmarkMakeValid/makevalidTestCases_#2_Four_columns_invalid_multipolygon-4     596480        141918        -76.21%
BenchmarkMakeValid/makevalidTestCases_#3_Square-4                                73202         33179         -54.67%
BenchmarkMakeValid/makevalidTestCases_#0_Triangle-4                              63537         25593         -59.72%

benchmark                                                                        old allocs     new allocs     delta
BenchmarkMakeValid/makevalidTestCases_#1_Four_Square_IO_OI-4                     1967           565            -71.28%
BenchmarkMakeValid/makevalidTestCases_#2_Four_columns_invalid_multipolygon-4     6969           1029           -85.23%
BenchmarkMakeValid/makevalidTestCases_#3_Square-4                                833            298            -64.23%
BenchmarkMakeValid/makevalidTestCases_#0_Triangle-4                              673            237            -64.78%
```

I've also created first patch to use slice-based allocation of QuadEdges, instead of pointers. 
https://github.com/olt/geom/commit/493534f8c6a59fccb7dabdce7fe266d3c1fef6d2
This improves performance and allocations even further. This requires more changes however as all MakeEdge calls now need to made from a new QuadEdgeSet/Collection (the current patch uses a global slice).

```
benchmark                                                                        old ns/op     new ns/op     delta
BenchmarkMakeValid/makevalidTestCases_#1_Four_Square_IO_OI-4                     188019        63991         -65.97%
BenchmarkMakeValid/makevalidTestCases_#2_Four_columns_invalid_multipolygon-4     596480        131656        -77.93%
BenchmarkMakeValid/makevalidTestCases_#3_Square-4                                73202         31633         -56.79%
BenchmarkMakeValid/makevalidTestCases_#0_Triangle-4                              63537         26882         -57.69%

benchmark                                                                        old allocs     new allocs     delta
BenchmarkMakeValid/makevalidTestCases_#1_Four_Square_IO_OI-4                     1967           451            -77.07%
BenchmarkMakeValid/makevalidTestCases_#2_Four_columns_invalid_multipolygon-4     6969           884            -87.32%
BenchmarkMakeValid/makevalidTestCases_#3_Square-4                                833            228            -72.63%
BenchmarkMakeValid/makevalidTestCases_#0_Triangle-4                              673            179            -73.40%
```

Another low hanging fruit is the `rtreego` implementation which is a generic RTree that supports multiple dimensions.  ~20% of all allocations are within this package. This could be optimized by providing an implementation that stores the `rect` coordinates in the nodes instead of pointers to Point for each dimension.