@gdey 2018-08-15
Great work! the rtreego is internal so that we can swap it out with an implementation that is more suited to our needs but just needed something to get to the point where we know the algo is working. I'm still seeing bugs in the Triangulation part, that I'm trying to work out.