@ARolek 2018-08-15
@olt nice! Thanks for the optimizations and the associated benchmarks. Looking forward to having these merged in.