@olt 2018-08-17
Another bunch improvements with an additional -13-15% ns/op and -30-50% allocs. Not building the RTree for small geometries is a huge boost.
I've integrated these changes into Tegola and the time for a level 13 OSM tile (with buildings, roads, etc.) went down from 1.6 to 0.4 seconds.