@ARolek 2018-08-20
@olt these optimizations are great! Thanks for helping with the performance research. Looking forward to having these implemented.