@gdey 2018-09-07
@olt I believe I got the bulk of the proposed change in here merged in. It may make sense to rebase the branch on master before looking at more areas to improve.