@gdey 2018-09-17
Going to close this branch. @olt When I'm done fixing the issues in the [gdey_delaunay](https://github.com/go-spatial/geom/tree/gdey_delaunay) branch could you take a look at it.